♻️ Refact: UserUpdateFacade의 프로필 이미지 업데이트 로직 수정
- 기존 프로필 이미지 삭제 조건 및 처리 로직 개선
- 프로필 이미지 생성 및 업데이트 로직 명확화
- 코드 가독성 및 유지보수성 향상

+++ src/main/java/com/sprint/mission/discodeit/facade/user/UserUpdateFacade.java
@@ -25,19 +25,20 @@ public class UserUpdateFacade {
         User user = userService.findById(userId);
         UUID profileId = null;
 
+        //이메일과 닉네임 부터 update(조건 맞지 않으면 바로 예외처리)
+        userService.update(userId, req);
 
+        //기존 프로필 사진 있으면 무조건 삭제
+        if(user.getProfileId() != null){binaryContentService.delete(user.getProfileId());}
+
+        //올라온 데이터가 있으면 무조건 만들어서 배정
         if(req.profileImage().data() != null){
-            //기존 사진이 있으면 삭제
-            if(user.getProfileId() != null){
-                binaryContentService.delete(user.getProfileId());
-            }
             BinaryContent profileImg = binaryContentService.create(
-                BinaryContentFactory.create(req.profileImage())
+                    BinaryContentFactory.create(req.profileImage())
             );
             profileId = profileImg.getId();
-
         }
-        userService.update(userId, req);
+
         userService.updateProfileImage(userId, profileId);
         userStatusService.updateByUserId(userId);
     }
